manualToolSample/requirements.txt
Django==3.2.25
psycopg2-binary==2.9.9
Pillow==10.4.0
django-extensions
matplotlib
pytest-django